move abc / fix timimg
